+++ src/Exception/Config/InvalidFileTypeException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PhpZone\PhpZone\Exception\Config;
+
+class InvalidFileTypeException extends \Exception
+{
+}