+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased][unreleased]
+### Added
+- Support for importing resources in config file.
+
 ### Changed
-- Registration of extensions via Symfony Dependency Injection Extension system
+- Registration of extensions via Symfony Dependency Injection Extension system.
 
 ## 0.1.0 - 2015-04-07
 ### Added